+++ sipx/examples/metrics.py
@@ -1,7 +1,12 @@
 import asyncio
 
 from sipx import SipCallError, SipUac
-from sipx.examples.common import account_settings, debug_wire, print_json
+from sipx.examples.common import (
+    account_settings,
+    debug_wire,
+    debug_wire_rtp,
+    print_json,
+)
 
 
 async def metrics() -> None:
@@ -18,7 +23,7 @@ async def metrics() -> None:
         local_host=s["local_host"],
         local_port=s["local_port"],
         timeout=s["timeout"],
-        event_hooks={"wire": [debug_wire]},
+        event_hooks={"wire": [debug_wire], "rtp": [debug_wire_rtp]},
     ) as uac:
         try:
             call = await asyncio.wait_for(